Add --force-template-json to diff
The `diff` command gives a complete change when the proposed and current
stack formats aren't the same. This is unhelpful when attempting to move
an existing stack (defined as JSON) into stack_master as a YAML
template.

This commit adds a `force_template_json` option to the StackDiffer, and
updates the `diff` command to take a `--force-template-json` flag to
compare the normalized format.